+++ R/specify.R
@@ -689,6 +689,7 @@ specify_all <- function(.tidy_iea_df){
 #' @param notations the notations used for row and column names. See `matsbyname::notation_vec()`. 
 #'                 Default is `list(IEATools::of_notation, IEATools::from_notation)`, 
 #'                 because both `IEATools::of_notation` and `IEATools::from_notation` can be used in the `Flow` column.
+#' @param production,resources See `IEATools::tpes_flows`.
 #'
 #' @return a de-specified version of `.df`
 #' 
@@ -701,12 +702,15 @@ specify_all <- function(.tidy_iea_df){
 #'   despecify_col(col = "Flow", despecified_col = "clean_Flow") %>% 
 #'   select(Flow, Product, E.dot, clean_Flow) %>% 
 #'   filter(endsWith(Flow, bracket_notation[["suff_end"]]))
-despecify_col <- function(.df, col, despecified_col, notations = list(IEATools::of_notation, IEATools::from_notation)) {
+despecify_col <- function(.df, col, despecified_col, 
+                          notations = list(IEATools::of_notation, IEATools::from_notation),
+                          production = IEATools::tpes_flows$production,
+                          resources = IEATools::tpes_flows$resources) {
   out <- .df %>% 
     dplyr::mutate(
       "{despecified_col}" := dplyr::case_when(
         # Change "Resources" back to "Production"
-        startsWith(.data[[col]], tpes_flows$resources) ~ tpes_flows$production, 
+        startsWith(.data[[col]], resources) ~ production, 
         TRUE ~ .data[[col]]
       )
     )