+++ NEWS.md
@@ -1,3 +1,7 @@
+* Fixed a bug in `despecify_col()`.
+  `resources` and `production` were not correctly using the `tpes_flows` object.
+
+
 # IEATools 0.1.21 (2020-05-14)
 
 * Bug fixes in `extend_to_useful()`.